refactor(mocker): share live scheduler loop
