+++ lib/mocker/src/scheduler/vllm/live.rs
@@ -1,9 +1,12 @@
 // SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-use std::collections::VecDeque;
+use std::time::Instant;
+
+#[cfg(all(test, feature = "kvbm-offload"))]
 use std::sync::{Arc, Mutex};
-use std::time::{Duration, Instant};
+#[cfg(all(test, feature = "kvbm-offload"))]
+use std::time::Duration;
 
 use tokio::sync::mpsc;
 use tokio_util::sync::CancellationToken;
@@ -11,11 +14,10 @@ use tokio_util::sync::CancellationToken;
 use crate::common::protocols::{
     DirectRequest, FpmPublisher, KvEventPublishers, MockEngineArgs, OutputSignal,
 };
-use crate::common::utils::sleep_until_precise;
 use crate::scheduler::{
-    AdmissionEvent, LiveBoundaryCore, LiveEffectsPublisher, SchedulerCommand,
+    AdmissionEvent, LiveBoundaryCore, LivePassExecution, LiveSchedulerState, SchedulerCommand,
     SchedulerCommandEffects, SchedulerCommandEnvelope, SchedulerHandle, SchedulerLifecycleEvent,
-    capture_deferred_kv_publish_sink, handoff_channel_capacity,
+    spawn_live_scheduler,
 };
 
 use super::core::VllmCore;
@@ -32,7 +34,6 @@ pub struct MockerMetrics {
     pub sglang_cache_hit_tokens: u64,
     pub sglang_cache_total_tokens: u64,
 }
-
 impl MockerMetrics {
     pub fn new(
         dp_rank: dynamo_kv_router::protocols::DpRank,
@@ -74,19 +75,7 @@ impl MockerMetrics {
 
 #[derive(Clone)]
 pub struct Scheduler {
-    request_tx: mpsc::UnboundedSender<DirectRequest>,
-    command_tx: mpsc::Sender<SchedulerCommandEnvelope>,
-    lifecycle_rx: Arc<Mutex<Option<mpsc::Receiver<SchedulerLifecycleEvent>>>>,
-    metrics_rx: tokio::sync::watch::Receiver<MockerMetrics>,
-    _cancel_guard: Arc<CancelGuard>,
-}
-
-struct CancelGuard(CancellationToken);
-
-impl Drop for CancelGuard {
-    fn drop(&mut self) {
-        self.0.cancel();
-    }
+    inner: LiveSchedulerState,
 }
 
 impl Scheduler {
@@ -138,407 +127,63 @@ impl Scheduler {
         admission_tx: Option<mpsc::UnboundedSender<AdmissionEvent>>,
         fpm_publisher: FpmPublisher,
     ) -> Self {
-        let (request_tx, mut request_rx) = mpsc::unbounded_channel::<DirectRequest>();
-        let control_capacity = handoff_channel_capacity(&args);
-        let (command_tx, mut command_rx) =
-            mpsc::channel::<SchedulerCommandEnvelope>(control_capacity);
-        let (lifecycle_tx, lifecycle_rx) =
-            mpsc::channel::<SchedulerLifecycleEvent>(control_capacity);
-        let total_blocks = args.num_gpu_blocks as u64;
-        let initial_metrics = MockerMetrics::new(dp_rank, 0, total_blocks);
-        let (metrics_tx, metrics_rx) = tokio::sync::watch::channel(initial_metrics);
-
-        let cancel_token = cancellation_token.unwrap_or_default();
-        let cancel_token_clone = cancel_token.clone();
-        let cancel_guard = Arc::new(CancelGuard(cancel_token));
-        let controls_enabled = args.is_prefill() || args.is_decode();
-
-        tokio::spawn(async move {
-            let (deferred_kv_events, buffering_publishers) = capture_deferred_kv_publish_sink(
-                !kv_event_publishers.is_empty(),
-                kv_event_publishers.raw_enabled(),
-            );
-            let mut core = VllmCore::new_with_sink(args, dp_rank, buffering_publishers);
-            let publisher = LiveEffectsPublisher::new(
-                output_tx,
-                admission_tx,
-                lifecycle_tx,
-                metrics_tx,
-                kv_event_publishers,
-                fpm_publisher,
-                deferred_kv_events,
-            );
-            #[cfg(feature = "kvbm-offload")]
-            if let Err(e) = core.init_offload_live().await {
-                tracing::error!("kvbm-offload live init failed: {e}");
-            }
-            // Wall-clock origin for this scheduler's simulated time. Drives
-            // `engine.tick(now_ms)` so the PS bandwidth models advance
-            // in real time across passes.
-            let scheduler_start = Instant::now();
-            let mut deferred_commands = VecDeque::new();
-
-            loop {
-                if !receive_until_schedulable(
-                    &mut core,
-                    &mut request_rx,
-                    &mut command_rx,
-                    &publisher,
-                    &scheduler_start,
-                    &cancel_token_clone,
-                    controls_enabled,
-                )
-                .await
-                {
-                    break;
-                }
-
-                let iteration_start = Instant::now();
-                let now_ms = scheduler_start.elapsed().as_secs_f64() * 1000.0;
-                let metrics_before = core.mocker_metrics();
-                let pass = core.execute_pass_internal(None, now_ms, None);
-                let mut pending = publisher.capture_pass(pass);
-                let total_time = std::time::Duration::from_secs_f64(
-                    (pending.end_ms() - now_ms).max(0.0) / 1000.0,
-                );
-                let zero_progress =
-                    total_time.is_zero() && !pending.made_progress_since(&metrics_before);
-                publisher.publish_pass_start(&mut pending);
-                if total_time > std::time::Duration::ZERO {
-                    let deadline = iteration_start + total_time;
-                    if controls_enabled {
-                        if !wait_for_pass_boundary(
-                            &mut core,
-                            &mut command_rx,
-                            &mut deferred_commands,
-                            &publisher,
-                            &scheduler_start,
-                            &cancel_token_clone,
-                            deadline,
-                        )
-                        .await
-                        {
-                            break;
-                        }
-                    } else {
-                        sleep_until_precise(deadline).await;
-                    }
-                }
-                publisher.publish_pass(&mut core, pending).await;
-                if controls_enabled {
-                    let mut command_processed = false;
-                    while let Some(command) = deferred_commands.pop_front() {
-                        command_processed = true;
-                        publisher
-                            .apply_command(
-                                &mut core,
-                                command,
-                                true,
-                                scheduler_elapsed_ms(&scheduler_start),
-                            )
-                            .await;
-                    }
-                    while let Ok(command) = command_rx.try_recv() {
-                        command_processed = true;
-                        publisher
-                            .apply_command(
-                                &mut core,
-                                command,
-                                true,
-                                scheduler_elapsed_ms(&scheduler_start),
-                            )
-                            .await;
-                    }
-                    let retry_progress = publisher
-                        .retry_destinations(&mut core, scheduler_elapsed_ms(&scheduler_start))
-                        .await;
-                    if zero_progress
-                        && !command_processed
-                        && !retry_progress
-                        && !wait_for_progress_wake(
-                            &mut core,
-                            &mut request_rx,
-                            &mut command_rx,
-                            &publisher,
-                            &scheduler_start,
-                            &cancel_token_clone,
-                            true,
-                        )
-                        .await
-                    {
-                        break;
-                    }
-                } else if zero_progress
-                    && !wait_for_progress_wake(
-                        &mut core,
-                        &mut request_rx,
-                        &mut command_rx,
-                        &publisher,
-                        &scheduler_start,
-                        &cancel_token_clone,
-                        false,
-                    )
-                    .await
-                {
-                    break;
-                }
-            }
-        });
-
         Self {
-            request_tx,
-            command_tx,
-            lifecycle_rx: Arc::new(Mutex::new(Some(lifecycle_rx))),
-            metrics_rx,
-            _cancel_guard: cancel_guard,
+            inner: spawn_live_scheduler(
+                args,
+                dp_rank,
+                output_tx,
+                kv_event_publishers,
+                cancellation_token,
+                admission_tx,
+                fpm_publisher,
+                VllmCore::new_with_sink,
+            ),
         }
     }
 }
 
 impl SchedulerHandle for Scheduler {
     fn receive(&self, request: DirectRequest) {
-        let _ = self.request_tx.send(request);
+        self.inner.receive(request);
     }
 
     fn request_sender(&self) -> mpsc::UnboundedSender<DirectRequest> {
-        self.request_tx.clone()
+        self.inner.request_sender()
     }
 
     fn metrics_receiver(&self) -> tokio::sync::watch::Receiver<MockerMetrics> {
-        self.metrics_rx.clone()
+        self.inner.metrics_receiver()
     }
 
     fn command_sender(&self) -> mpsc::Sender<SchedulerCommandEnvelope> {
-        self.command_tx.clone()
+        self.inner.command_sender()
     }
 
     fn take_lifecycle_receiver(&mut self) -> Option<mpsc::Receiver<SchedulerLifecycleEvent>> {
-        self.lifecycle_rx
-            .lock()
-            .expect("scheduler lifecycle receiver mutex poisoned")
-            .take()
+        self.inner.take_lifecycle_receiver()
     }
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn receive_until_schedulable(
-    core: &mut VllmCore,
-    request_rx: &mut mpsc::UnboundedReceiver<DirectRequest>,
-    command_rx: &mut mpsc::Receiver<SchedulerCommandEnvelope>,
-    publisher: &LiveEffectsPublisher,
-    scheduler_start: &Instant,
-    cancel_token: &CancellationToken,
-    controls_enabled: bool,
-) -> bool {
-    if !controls_enabled {
-        if cancel_token.is_cancelled() {
-            return false;
-        }
-        if core.is_empty() {
-            tokio::select! {
-                biased;
-                _ = cancel_token.cancelled() => return false,
-                request = request_rx.recv() => {
-                    let Some(request) = request else {
-                        return false;
-                    };
-                    core.receive(request);
-                }
-            }
-        }
-        while let Ok(request) = request_rx.try_recv() {
-            core.receive(request);
-        }
-        return true;
-    }
-
-    while core.is_empty() {
-        #[cfg(feature = "kvbm-offload")]
-        let internal_deadline_ms = core.earliest_offload_deadline();
-        #[cfg(not(feature = "kvbm-offload"))]
-        let internal_deadline_ms = None;
-        let internal_deadline = wait_for_internal_deadline(scheduler_start, internal_deadline_ms);
-        tokio::pin!(internal_deadline);
-        tokio::select! {
-            biased;
-            _ = cancel_token.cancelled() => return false,
-            command = command_rx.recv() => {
-                let Some(command) = command else {
-                    return false;
-                };
-                publisher
-                    .apply_command(core, command, true, scheduler_elapsed_ms(scheduler_start))
-                    .await;
-            }
-            result = request_rx.recv() => {
-                let Some(request) = result else {
-                    return false;
-                };
-                core.receive(request);
-            }
-            _ = &mut internal_deadline, if internal_deadline_ms.is_some() => {
-                #[cfg(feature = "kvbm-offload")]
-                {
-                    let now_ms = scheduler_elapsed_ms(scheduler_start)
-                        .max(internal_deadline_ms.expect("armed offload deadline"));
-                    publisher.advance_offload(core, now_ms, true).await;
-                }
-            }
-        }
-    }
-
-    while let Ok(command) = command_rx.try_recv() {
-        publisher
-            .apply_command(core, command, true, scheduler_elapsed_ms(scheduler_start))
-            .await;
-    }
-    while let Ok(request) = request_rx.try_recv() {
-        core.receive(request);
-    }
-
-    true
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn wait_for_pass_boundary(
-    core: &mut VllmCore,
-    command_rx: &mut mpsc::Receiver<SchedulerCommandEnvelope>,
-    deferred_commands: &mut VecDeque<SchedulerCommandEnvelope>,
-    publisher: &LiveEffectsPublisher,
-    scheduler_start: &Instant,
-    cancel_token: &CancellationToken,
-    deadline: Instant,
-) -> bool {
-    let sleep = sleep_until_precise(deadline);
-    tokio::pin!(sleep);
-    let mut accept_commands = true;
-    loop {
-        #[cfg(feature = "kvbm-offload")]
-        let internal_deadline_ms = core.earliest_offload_deadline();
-        #[cfg(not(feature = "kvbm-offload"))]
-        let internal_deadline_ms = None;
-        let internal_deadline = wait_for_internal_deadline(scheduler_start, internal_deadline_ms);
-        tokio::pin!(internal_deadline);
-        tokio::select! {
-            biased;
-            _ = cancel_token.cancelled() => return false,
-            _ = &mut sleep => return true,
-            _ = &mut internal_deadline, if internal_deadline_ms.is_some() => {
-                #[cfg(feature = "kvbm-offload")]
-                {
-                    let now_ms = scheduler_elapsed_ms(scheduler_start)
-                        .max(internal_deadline_ms.expect("armed offload deadline"));
-                    publisher.advance_offload(core, now_ms, false).await;
-                    debug_assert!(
-                        core.earliest_offload_deadline()
-                            .is_none_or(|next| next > now_ms),
-                        "offload tick left an already-due deadline armed"
-                    );
-                }
-            }
-            command = command_rx.recv(), if accept_commands => {
-                let Some(command) = command else {
-                    return false;
-                };
-                if command_can_apply_during_pass(&command.command) {
-                    publisher
-                        .apply_command(
-                            core,
-                            command,
-                            false,
-                            scheduler_elapsed_ms(scheduler_start),
-                        )
-                        .await;
-                } else {
-                    deferred_commands.push_back(command);
-                    accept_commands = false;
-                }
-            }
-        }
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn wait_for_progress_wake(
-    core: &mut VllmCore,
-    request_rx: &mut mpsc::UnboundedReceiver<DirectRequest>,
-    command_rx: &mut mpsc::Receiver<SchedulerCommandEnvelope>,
-    publisher: &LiveEffectsPublisher,
-    scheduler_start: &Instant,
-    cancel_token: &CancellationToken,
-    controls_enabled: bool,
-) -> bool {
-    #[cfg(feature = "kvbm-offload")]
-    let internal_deadline_ms = core.earliest_offload_deadline();
-    #[cfg(not(feature = "kvbm-offload"))]
-    let internal_deadline_ms = None;
-
-    let internal_deadline = wait_for_internal_deadline(scheduler_start, internal_deadline_ms);
-    tokio::pin!(internal_deadline);
-    if controls_enabled {
-        tokio::select! {
-            biased;
-            _ = cancel_token.cancelled() => false,
-            command = command_rx.recv() => {
-                let Some(command) = command else {
-                    return false;
-                };
-                publisher
-                    .apply_command(core, command, true, scheduler_elapsed_ms(scheduler_start))
-                    .await;
-                true
-            }
-            request = request_rx.recv() => {
-                let Some(request) = request else {
-                    return false;
-                };
-                core.receive(request);
-                true
-            }
-            _ = &mut internal_deadline => true,
-        }
-    } else {
-        tokio::select! {
-            biased;
-            _ = cancel_token.cancelled() => false,
-            request = request_rx.recv() => {
-                let Some(request) = request else {
-                    return false;
-                };
-                core.receive(request);
-                true
-            }
-            _ = &mut internal_deadline => true,
-        }
-    }
-}
-
-async fn wait_for_internal_deadline(scheduler_start: &Instant, deadline_ms: Option<f64>) {
-    let Some(deadline_ms) = deadline_ms else {
-        std::future::pending::<()>().await;
-        return;
-    };
-    let deadline = *scheduler_start + Duration::from_secs_f64(deadline_ms.max(0.0) / 1000.0);
-    let wake_at = if deadline <= Instant::now() {
-        Instant::now() + Duration::from_millis(1)
-    } else {
-        deadline
-    };
-    sleep_until_precise(wake_at).await;
-}
-
-fn scheduler_elapsed_ms(scheduler_start: &Instant) -> f64 {
-    scheduler_start.elapsed().as_secs_f64() * 1000.0
-}
-
-fn command_can_apply_during_pass(command: &SchedulerCommand) -> bool {
-    matches!(
-        command,
-        SchedulerCommand::SubmitHandoffPrefill { .. } | SchedulerCommand::ReserveDestination { .. }
-    )
 }
 
 impl LiveBoundaryCore for VllmCore {
+    fn initialize_live(
+        &mut self,
+    ) -> std::pin::Pin<Box<dyn std::future::Future<Output = ()> + Send + '_>> {
+        Box::pin(async move {
+            #[cfg(feature = "kvbm-offload")]
+            if let Err(error) = self.init_offload_live().await {
+                tracing::error!("kvbm-offload live init failed: {error}");
+            }
+        })
+    }
+
+    fn live_is_empty(&self) -> bool {
+        self.is_empty()
+    }
+
+    fn receive_live_request(&mut self, request: DirectRequest) {
+        self.receive(request);
+    }
+
     fn apply_live_command(
         &mut self,
         command: SchedulerCommand,
@@ -558,6 +203,26 @@ impl LiveBoundaryCore for VllmCore {
 
     fn pass_boundary_metrics(&self, _pass_metrics: MockerMetrics) -> MockerMetrics {
         self.mocker_metrics()
+    }
+
+    fn live_internal_deadline_ms(&self) -> Option<f64> {
+        #[cfg(feature = "kvbm-offload")]
+        {
+            self.earliest_offload_deadline()
+        }
+        #[cfg(not(feature = "kvbm-offload"))]
+        {
+            None
+        }
+    }
+
+    fn execute_live_pass(&mut self, scheduler_start: &Instant) -> LivePassExecution {
+        // Wall-clock elapsed time drives the PS bandwidth models across
+        // vLLM passes; SGLang reports a duration directly from each pass.
+        let now_ms = scheduler_start.elapsed().as_secs_f64() * 1000.0;
+        let pass = self.execute_pass_internal(None, now_ms, None);
+        let duration = std::time::Duration::from_secs_f64((pass.end_ms - now_ms).max(0.0) / 1000.0);
+        LivePassExecution { pass, duration }
     }
 
     fn output_delivery_failed(&mut self, signals: Vec<OutputSignal>) {
